requirements.txt
Pillow
google-cloud-aiplatform
google-cloud-texttospeech